+++ src/app/station-board/search/custom-stop-event-xml-popover/custom-stop-event-xml-popover.component.ts
@@ -4,12 +4,15 @@ import { UserTripService } from 'src/app/shared/services/user-trip.service';
 import * as OJP from 'ojp-sdk';
 
 import { LanguageService } from '../../../shared/services/language.service';
+import { APP_STAGE, DEFAULT_APP_STAGE } from '../../../config/constants';
 
 @Component({
   selector: 'custom-stop-event-popover',
   templateUrl: './custom-stop-event-xml-popover.component.html',
 })
 export class CustomStopEventXMLPopoverComponent {
+  public appStage: APP_STAGE;
+  
   public customRequestXMLs: string
   public customResponseXMLs: string
 
@@ -19,6 +22,8 @@ export class CustomStopEventXMLPopoverComponent {
   @Output() customResponseSaved = new EventEmitter<string>()
 
   constructor(private userTripService: UserTripService, private languageService: LanguageService) {
+    this.appStage = DEFAULT_APP_STAGE;
+
     this.customRequestXMLs = '... wait'
     this.customResponseXMLs = 'Paste custom OJP StopEventRequest Response XML here...'
 
@@ -26,7 +31,7 @@ export class CustomStopEventXMLPopoverComponent {
   }
 
   public async parseCustomRequestXML() {
-    const sdk = this.userTripService.createOJP_SDK_Instance(this.languageService.language);
+    const sdk = this.userTripService.createOJP_SDK_Instance(this.languageService.language, this.appStage);
     const request = sdk.requests.StopEventRequest.initWithRequestMock(this.customRequestXMLs);
     
     this.isRunningRequest = true;